src/requirements.txt
IMDbPY
mysqlclient
MySQL-python
